Extend support of route creation when tgw is not created
The aws_route resource is dependent on the tgw being created within the
module. If the tgw is not created in the module it is not possible to
update routes. By using the transit_gateway_id output available from the
vpc attachment we can support a tgw created outside of the module.